feature: add unified localization file, switcher
The default behavior of ASP.NET is to pull localization strings from a
resx file for each controller in the application. To simplify
collaboration/editing of these resource files, the behavior was modified
to allow injection of localized strings from a unified resource file,
stored in application source at Resources/LocalizationData.culture.resx

Application code to allow changing the languge application-wide via a
cookie was added. A language selector was placed in the footer.

The culture may also be specified by the browser in the Content-Language
header, or with the query string ?culture=<culture> on the end of each
request. The cookie approach allows users to store a language for their
session regardless of browser preference.j