design: second design clean up pass
Made home page look more homey -- fixed sizing and alignment of title
objects, added a box which can be filled with a link to the most recent
article, and added a placeholder at the bottom for more information
about the project and team.

Cleaned up the Articles Index page to be consistent with alignment and
sizing on the Home page and the Dashboard.

Added some paint to the article view page.