participate: minimal response tracking
adds a "never-expiring" cookie to track whether a user has already
responded to a survey, and allows for answering of multiple surveys.
needs some code cleanup (DRY, make it easier to access available
surveys) and also some UX/view cleanup.

in the future this should be strengthened by trying to re-associate a
cookie based on IP address/browser fingerprint if a cache is cleared.